PremiumCare/requirements.txt
# MedPlatform — Pinned Requirements
# Generated: 2026-03-18
# Update with: pip install --upgrade <package> then pin new version

# ── Core framework ─────────────────────────────────────────────
flask==3.0.3
flask-bcrypt==1.0.1
werkzeug==3.0.3

# ── Database ───────────────────────────────────────────────────
psycopg2-binary==2.9.9

# ── AI providers ──────────────────────────────────────────────
groq==0.9.0
google-generativeai==0.7.2

# ── Security ──────────────────────────────────────────────────
argon2-cffi==23.1.0
cryptography==42.0.8
pyotp==2.9.0
bleach==6.1.0
flask-talisman==1.1.0
flask-wtf==1.2.1
qrcode[pil]==7.4.2

# ── Sessions & caching ────────────────────────────────────────
redis==5.0.4
flask-session==0.8.0

# ── Scheduler ─────────────────────────────────────────────────
apscheduler==3.10.4

# ── Utilities ─────────────────────────────────────────────────
python-dotenv==1.0.1
Pillow==10.3.0
flask-limiter==3.5.0
limits==3.12.0
